// gen/ts/src/common/types.ts
export enum SpaceType {
  SPACE_TYPE_UNSPECIFIED = "SPACE_TYPE_UNSPECIFIED",
  BUILDING = "BUILDING",
  FLOOR = "FLOOR",
  APARTMENT = "APARTMENT",
  ROOM = "ROOM",
  COMMON_AREA = "COMMON_AREA",
  UNRECOGNIZED = "UNRECOGNIZED",
}