// gen/ts/src/space/space.ts
export interface Space {
  spaceId?: SpaceId | undefined;
  spaceType?: SpaceType | undefined;
  name?: string | undefined;
  description?:
    | string
    | undefined;
  /** meaningful for FLOOR type */
  floor?:
    | number
    | undefined;
  /** null for top-level Building */
  parentSpaceId?: SpaceId | undefined;
  subSpaceIds?: SpaceId[] | undefined;
  deviceIds?:
    | DeviceId[]
    | undefined;
  /** only for APARTMENT */
  residentUserId?: UserId | undefined;
  tenantId?: TenantId | undefined;
  gatewayId?: GatewayId | undefined;
}